# swagger-ui/openapi3.yaml
components:
  requestBodies:
    PSQLRequest:
      content:
        application/json:
          schema:
            properties:
              id:
                minLength: 1
                type: string
              requestType:
                minLength: 1
                type: string
              table:
                minLength: 1
                type: string
      description: Request used for fetching postgres data
      required: true
  responses:
    PSQLResponse:
      content:
        application/json:
          schema:
            nullable: true
            properties:
              lastInsertID:
                type: string
              message:
                $ref: '#/components/schemas/Message'
              rowsAffected:
                nullable: true
                type: string
      description: Response with postgres data
  schemas:
    ErrorLog:
      properties:
        query:
          nullable: true
          type: string
        rootCause:
          nullable: true
          type: string
        scope:
          nullable: true
          type: string
        status:
          nullable: true
          type: string
        trace:
          nullable: true
          type: string
      type: object
    ErrorLogs:
      items:
        $ref: '#/components/schemas/ErrorLog'
      type: array
    Message:
      properties:
        count:
          nullable: true
          type: string
        errorLog:
          $ref: '#/components/schemas/ErrorLogs'
        hostName:
          nullable: true
          type: string
        status:
          nullable: true
          type: string
        timeTaken:
          nullable: true
          type: string
      type: object
info:
  contact:
    url: https://github.com/CalebTracey/api-template
  description: REST API used for accessing postres database
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go REST API template
  version: 0.0.0
openapi: 3.0.0
paths:
  /add:
    description: Add Data
    post:
      operationId: AddToDatabase
      requestBody:
        $ref: '#/components/requestBodies/PSQLRequest'
      responses:
        "201":
          $ref: '#/components/responses/PSQLResponse'
        "400":
          $ref: '#/components/responses/PSQLResponse'
        "500":
          $ref: '#/components/responses/PSQLResponse'
servers:
- description: Local development
  url: http://0.0.0.0:6080